fix : Suspense build 이슈 수정
